Fix parallel stereoscopic rendering to better suit 3D TVs.
The old parallel method assumed you'd use the parallel viewing method
on a regular screen. It used a 8px border to separate the frames, and
forced a low output resolution.

The changes remove the 8px border, which is not needed when the frames
are output on a 3D TV, and it also greatly complicates things.