Make the "dump prepass" key (d) always available; show the prepass on-screen.
Previously the dump prepass option was only a debug feature. When you pressed
'd', it would generate a couple of .bmp files in the current dir, containing
the prepass and the final render for the current frame. This is different now -
the 'd' key is a toggle, and, when on, the prepass is shown onscreen, instead
of the shaded rendering.